chore(refactor): Move specific interfaces/utils
