+++ src/app/index.ts
@@ -1,1 +1,3 @@
 export * from './firebase.app.module';
+export * from './interfaces';
+export * from './tokens';
