+++ src/app/interfaces.ts
@@ -1,0 +1,7 @@
+export interface FirebaseAppConfig {
+  apiKey?: string;
+  authDomain?: string;
+  databaseURL?: string;
+  storageBucket?: string;
+  messagingSenderId?: string;
+}
