+++ src/app/tokens.ts
@@ -1,0 +1,2 @@
+import { OpaqueToken } from '@angular/core';
+export const FirebaseAppName = new OpaqueToken('FirebaseAppName');
