+++ src/auth/auth.ts
@@ -1,8 +1,8 @@
 import * as firebase from 'firebase/app';
-import * as utils from '../utils';
+import { ZoneScheduler } from '../utils';
 import 'firebase/auth';
 import { Injectable, NgZone } from '@angular/core';
-import { Auth } from '../interfaces';
+import { Auth } from './interfaces';
 import { Observable } from 'rxjs/Observable';
 import { Observer } from 'rxjs/Observer';
 import { observeOn } from 'rxjs/operator/observeOn';
@@ -15,7 +15,7 @@ export class AngularFireAuth {
    * Firebase Auth instance
    */
   auth: firebase.auth.Auth;
-  
+
   /**
    * Observable of authentication state
    */
@@ -41,5 +41,5 @@ export function FirebaseAuthStateObservable(app: FirebaseApp): Observable<fireba
       () => observer.complete()
     );
   });
-  return observeOn.call(authState, new utils.ZoneScheduler(Zone.current));
+  return observeOn.call(authState, new ZoneScheduler(Zone.current));
 }
