+++ src/auth/index.ts
@@ -1,2 +1,3 @@
 export * from './auth';
 export * from './auth.module';
+export * from './interfaces';
