+++ src/auth/interfaces.ts
@@ -1,0 +1,2 @@
+import * as firebase from 'firebase/app';
+export type Auth = firebase.auth.Auth;
