+++ src/database/firebase_list_observable.ts
@@ -4,8 +4,9 @@ import { Subscriber } from 'rxjs/Subscriber';
 import { Subscription } from 'rxjs/Subscription';
 import * as firebase from 'firebase/app';
 import 'firebase/database';
-import * as utils from '../utils';
-import { AFUnwrappedDataSnapshot, FirebaseOperationCases, QueryReference, DatabaseSnapshot, DatabaseReference } from '../interfaces';
+import { isString } from '../utils';
+import { isFirebaseRef, isFirebaseDataSnapshot, isAFUnwrappedSnapshot } from './utils';
+import { AFUnwrappedDataSnapshot, FirebaseOperationCases, QueryReference, DatabaseSnapshot, DatabaseReference } from './interfaces';
 
 export type FirebaseOperation = string | firebase.database.Reference | firebase.database.DataSnapshot | AFUnwrappedDataSnapshot;
 
@@ -51,13 +52,13 @@ export class FirebaseListObservable<T> extends Observable<T> {
   }
 
   _checkOperationCases(item: FirebaseOperation, cases: FirebaseOperationCases) : firebase.Promise<void> {
-    if (utils.isString(item)) {
+    if (isString(item)) {
       return cases.stringCase();
-    } else if (utils.isFirebaseRef(item)) {
+    } else if (isFirebaseRef(item)) {
       return cases.firebaseCase();
-    } else if (utils.isFirebaseDataSnapshot(item)) {
+    } else if (isFirebaseDataSnapshot(item)) {
       return cases.snapshotCase();
-    } else if (utils.isAFUnwrappedSnapshot(item)) {
+    } else if (isAFUnwrappedSnapshot(item)) {
       return cases.unwrappedSnapshotCase()
     }
     throw new Error(`Method requires a key, snapshot, reference, or unwrapped snapshot. Got: ${typeof item}`);
