+++ src/database/firebase_object_factory.ts
@@ -3,8 +3,9 @@ import { Observer } from 'rxjs/Observer';
 import { observeOn } from 'rxjs/operator/observeOn';
 import * as firebase from 'firebase/app';
 import 'firebase/database';
-import * as utils from '../utils';
-import { FirebaseObjectFactoryOpts, PathReference, DatabaseReference } from '../interfaces';
+import { isAbsoluteUrl, ZoneScheduler } from '../utils';
+import { checkForUrlOrFirebaseRef, unwrapMapFn } from './utils';
+import { FirebaseObjectFactoryOpts, PathReference, DatabaseReference } from './interfaces';
 
 export function FirebaseObjectFactory (
   pathRef: PathReference,
@@ -12,21 +13,21 @@ export function FirebaseObjectFactory (
 
   let ref: DatabaseReference;
 
-  utils.checkForUrlOrFirebaseRef(pathRef, {
+  checkForUrlOrFirebaseRef(pathRef, {
     isUrl: () => {
       const path = pathRef as string;
-      if(utils.isAbsoluteUrl(path)) {
+      if(isAbsoluteUrl(path)) {
         ref = firebase.database().refFromURL(path)
       } else {
         ref = firebase.database().ref(path);
-      }      
+      }
     },
     isRef: () => ref = <DatabaseReference>pathRef
   });
 
   const objectObservable = new FirebaseObjectObservable((obs: Observer<any>) => {
     let fn = ref.on('value', (snapshot: firebase.database.DataSnapshot) => {
-      obs.next(preserveSnapshot ? snapshot : utils.unwrapMapFn(snapshot))
+      obs.next(preserveSnapshot ? snapshot : unwrapMapFn(snapshot))
     }, err => {
       if (err) { obs.error(err); obs.complete(); }
     });
@@ -35,5 +36,5 @@ export function FirebaseObjectFactory (
   }, ref);
 
   // TODO: should be in the subscription zone instead
-  return observeOn.call(objectObservable, new utils.ZoneScheduler(Zone.current));
+  return observeOn.call(objectObservable, new ZoneScheduler(Zone.current));
 }
