+++ src/database/index.ts
@@ -5,3 +5,5 @@ export * from './firebase_object_factory';
 export * from './firebase_object_observable';
 export * from './query_observable';
 export * from './database.module';
+export * from './interfaces';
+export * from './utils';
