+++ src/database/query_observable.spec.ts
@@ -1,7 +1,7 @@
 import { Observable } from 'rxjs/Observable';
 import { Observer } from 'rxjs/Observer';
 import { Subject } from 'rxjs/Subject';
-import { Query, ScalarQuery } from '../interfaces';
+import { Query, ScalarQuery } from './interfaces';
 import { getOrderObservables, observeQuery } from './index';
 
 function scalarQueryTest(query: ScalarQuery, done: any) {
