--- src/tokens.ts
@@ -1,3 +1,0 @@
-import {OpaqueToken} from '@angular/core';
-export const FirebaseApp = new OpaqueToken('FirebaseApp')
-export const FirebaseAppName = new OpaqueToken('FirebaseAppName');
